pyproject.toml
[tool.poetry]
name = "vertibayespython"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{ include = "com", from = "src" }]

[tool.poetry.dependencies]
python = "^3.7.1"
pgmpy = "^0.1.16"
pandas = "^1.3.5"
requests = "^2.25.1"
vantage6-client = "^3.0.4"
setuptools = "^59.6.0"
wheel = "^0.37.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
vantage6-client = "^3.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
